+++ PhpSerializerNET/PhpSerializiationOptions.cs
@@ -3,6 +3,7 @@
   License, v. 2.0. If a copy of the MPL was not distributed with this
   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 **/
+using System;
 
 namespace PhpSerializerNET {
 
@@ -15,6 +16,12 @@ namespace PhpSerializerNET {
 		/// Default false.
 		/// </summary>
 		public bool ThrowOnCircularReferences { get; set; } = false;
+
+		/// <summary>
+		/// Whether to serialize enums as numeric values (integers) or using their string representation 
+		/// (via <see cref="Enum.ToString()"/>)
+		/// </summary>
+		public bool NumericEnums { get; set; } = true;
 
 		internal static PhpSerializiationOptions DefaultOptions = new();
 	}